Add CSV parser tests
