+++ CONTRIBUTING.md
@@ -1,0 +1,18 @@
+# Contributing
+
+## Running tests
+
+This project uses the Node.js built-in test runner.
+To execute the test suite, run:
+
+```bash
+npm test
+```
+
+or directly:
+
+```bash
+node --test
+```
+
+The tests live in the `tests/` directory and cover CSV parsing edge cases.
