+++ timeline.html
@@ -73,6 +73,7 @@
 </footer>
 
 <script src="https://cdn.jsdelivr.net/npm/dompurify@3.0.3/dist/purify.min.js"></script>
+<script src="scripts/parseCSV.js"></script>
 <script src="scripts/timeline.js"></script>
 </body>
 </html>
